requirements.txt
datasets==2.21.0
pandas==2.2.2
matplotlib==3.9.1
scikit-image==0.24.0
torch==2.3.1
fastapi[standard]==0.111.0
seaborn==0.13.2
dependency-injector==4.41.0
python-dotenv==1.0.1
uvicorn==0.30.1
gdown==5.2.0
sphinx==8.0.2
sphinx_rtd_theme==2.0.0
sphinx-autodoc-typehints==2.4.4
sphinxcontrib-napoleon==0.7